refactor(Template): Template 페이지 폴더링
